Add event router with demo interactive messages
This shows how a router can sort messages for triggers (like "create
project" or "create file") and then handle that event by sending an
message with interactive components to begin the project instantiation
cycle.

Of course this is just a demo; in prod the list of project and file
templates would be built dynamically from the contents of the
lsst/templates repo, using templatekit to inspect it.