Handle a templatebot_file_dialog submission event
This adds routing to handle a dialog_submission event. As a
proof-of-concept, the handler posts the submitted content as a plain
text file into the channel (using the Slack files.upload method). This
will be a good way of providing the outputs from file/snippet generation
to users (they can just copy and paste the content into whatever editor
they're using).