Handle cases where a file template has no vars
If a file template doesn't have any configurability (like the GPLv3
license), then we need to short-cut the normal dialog interaction
pattern, and instead just respond with the template's content.

This required a bit of refactoring in the file_select handler to detect
that type of template, and refactoring in teh file_dialog_submission
handler to permit an API-based entry into template rendering and file
upload.