Add Avro schemas for templatebot events
These schemas define the messages that templatebot and its
organization-specific helper microservices pass to each other to setup,
render, and configure repositories built from templates.

These schemas take into consideration a possible retry algorithm
(timestamp of the initial message and retry count) and the possibility
of a Slack conversation related to the template creation.

Metadata about the template (name, variables, etc) are passed around
with each message so that all of the state is contained by the Kafka
messages.