Implement handle_project_render
This handler is triggered by the router with a templatebot-render_ready
event in SQuaRE Events (Kafka). The handler simply uses cookiecutter to
render the template, initialize the Git repository (using GitPython),
and then pushes that first commit to GitHub.

This adds GitPython as a new dependency.

There are a few todos still in this handler:

- Use the Slack API to get the user's name and email address for the
  commit.
- Send a confirmation or error message via Slack.