handle_project_render posts feedback to Slack
When the GitHub repo is populated, the handler sends a threaded message
over Slack. Likewise, if there's an error is also lets the user know.

The idea is that we should give enough feedback to the user over Slack
so they know what's happening, and to not have to wait unnecessarily if
there's a failure. Most debugging, though, should happen through
logging.