project dialog sub. handler updates original msg
When the project dialog submission handler receives a dialog_submission
event, it now replaces the original message (the one containing the
template selection menu) rather than posting a new message. This is made
possible by inserting the timestamp of the original message in the
`state` of the dialog.

Overall, this habit of replacing messages rather than adding new
messages makes for a cleaner Slack timeline.