Add _refresh_checkout method
The purpose of this method is to run idempotently whenever an existing
clone is being accessed based on a branch or tag symbolic git ref.

This method fetch's the cloned repo's origin and checks if the HEAD SHA
of the origin remote matches the cloned on. If it doesn't match, the
repo is re-cloned.