Upgrade to current faststream
The main hangup on faststream 0.5 was an apparent type issue with the
aiokafka KafkaMessage record. We'll ignore that typing issue to see if
its read.

The other issue is a forward type reference issue with pydantic; this is
solved by taking the future annotations out of the kafka handler module.