Add TemplateVariablesModal
This modal is derived from the base SlackModalView with a customized
constructor classmethod that creates the modal based on a TemplateKit
template. We're using the original templatebot recipe for creating
fields based on the templatekit.yaml definition for a template. Modals
provide a lot more types of inputs and for additional content in modals,
so we can improve this over time.

The private_metadata send with the modal is a pydantic model as well. It
includes all the metadata needed to for templatebot to pick up the
interaction and create a project from the template.

These tests of the modal construction use a mock template repository.
We're embedding it directly in this repo so we don't have a live
dependency on a real templates repo like
https://github.com/lsst/templates