Get GitHub repo setup variables from templates
Different templates have different methods for determining the GitHub
repo name and organization. Initially we do this with a series of cases,
but we might want to refactor this or make it more systematic to better
allow new templates to reliably set their intended github repo without
needing to modify this code.